Unify xgal ingestion stages
