+++ txpipe/ingest/__init__.py
@@ -1,5 +1,5 @@
 from .dp02 import TXIngestDataPreview02
-from .mocks import TXCosmoDC2Mock, TXGaussianSimsMock
+from .mocks import TXIngestRomanRubin, TXIngestSkySim, TXIngestCosmoDC2, TXIngestBuzzard, TXSimpleMock, TXMockTruthPZ
 from .gcr import TXMetacalGCRInput, TXIngestStars
 from .redmagic import TXIngestRedmagic
 from .ssi import (
@@ -11,4 +11,3 @@ from .ssi import (
 )
 from .dp1 import TXIngestDataPreview1
 from .legacy import TXIngestDESY3Gold
-from .skysim import TXIngestRomanRubin
